Move all task_list items components to assessment directory
