+++ app/views/planning_applications/assessment/tasks/_assess_against_legislation.html.erb
@@ -5,7 +5,7 @@
   <ul class="app-task-list__items" id="assess-policy-section">
     <% if @planning_application.application_type.name == "planning_permission" %>
         <%= render(
-            TaskListItems::LocalPolicyComponent.new(
+            TaskListItems::Assessment::LocalPolicyComponent.new(
               planning_application: @planning_application
             )
           ) %>
@@ -19,7 +19,7 @@
       <% else %>
         <% @planning_application.policy_classes.each do |policy_class| %>
           <%= render(
-            TaskListItems::PolicyClassComponent.new(
+            TaskListItems::Assessment::PolicyClassComponent.new(
               policy_class: policy_class,
               planning_application: @planning_application
             )
